Overwrite `preprocessor` and `ranlib`
Conda sets `preprocessor`, which we don't need or use. So go ahead and
drop it. Distutils sets `preprocessor` to `None` anyways. Meaning
Distutils can go on without it.

Also set `ranlib` to `None`. Both Conda and Distutils don't use it.
Still better safe than sorry. So we turn it off. We can always add it if
we need it later.